Refactor DTO naming support
